+++ code/modules/client/client_color.dm
@@ -2,7 +2,7 @@
 	var/client_color = "" //Any client.color-valid value
 	var/priority = 1
 	var/override = FALSE //If set to override we will stop multiplying the moment we get here. NOTE: Priority remains, if your override is on position 4, the other 3 will still have a say.
-
+	var/disability = FALSE
 
 /mob
 	var/list/client_colors = list()
@@ -22,9 +22,10 @@
 			return TRUE
 	return FALSE
 
-/mob/proc/add_client_color(color_type)
+/mob/proc/add_client_color(color_type, set_disability = FALSE)
 	if(!has_client_color(color_type))
 		var/datum/client_color/CC = new color_type()
+		CC.disability = set_disability
 		client_colors |= CC
 		sortTim(client_colors, /proc/cmp_clientcolor_priority)
 		update_client_color()
@@ -35,7 +36,7 @@
 	color_type - a typepath (subtyped from /datum/client_color)
 	returns true if instance was found, false otherwise
 */
-/mob/proc/remove_client_color(color_type)
+/mob/proc/remove_client_color(color_type, remove_disability = FALSE)
 	if(!ispath(/datum/client_color))
 		return FALSE
 
@@ -43,6 +44,8 @@
 	for(var/cc in client_colors)
 		var/datum/client_color/CC = cc
 		if(CC.type == color_type)
+			if(CC.disability && !remove_disability)
+				continue
 			result = TRUE
 			client_colors -= CC
 			qdel(CC)
@@ -93,7 +96,7 @@
 	priority = 300
 
 /datum/client_color/vaurca
-	client_color = list(0, 0, 0.297, 0, 0, 0.739, 0, 0, 0.100)
+	client_color = list(0.793, 0.297, 0, 0.297, 0.539, 0, 0, 0, 0)
 	priority = 100
 
 /datum/client_color/deuteranopia
@@ -109,7 +112,7 @@
 	priority = 100
 
 /datum/client_color/berserk
-	client_color = "#AF111C"
+	client_color = list(0.793, 0.4, 0.4, 0.793, 0.4, 0.4, 0, 0, 0)
 	priority = INFINITY //This effect sort of exists on its own you /have/ to be seeing RED
 	override = TRUE //Because multiplying this will inevitably fail
